refactor(notifications): reorganize notification utility functions into a dedicated utils module
