+++ backend/app/services/notifications/utils/__init__.py
@@ -1,0 +1,15 @@
+"""Notification utility functions."""
+
+from app.services.notifications.utils.content import (
+    build_eval_completion_payload,
+    build_eval_results_link,
+    format_completed_at,
+    notification_type_for_status,
+)
+
+__all__ = [
+    "build_eval_completion_payload",
+    "build_eval_results_link",
+    "format_completed_at",
+    "notification_type_for_status",
+]
